backend/requirements.txt
fastapi
uvicorn
pdf2image
pytesseract
paddleocr
sentence-transformers
faiss-cpu
python-multipart
requests 
together
python-dotenv
Flask 
streamlit 